Pruebas de Usuario y Vehiculo y algo de documentacion

+++ src/co/udea/iw/dao/ClienteDAO.java
@@ -4,12 +4,16 @@ import java.util.List;
 
 import co.udea.iw.dto.Cliente;
 import co.udea.iw.exception.IWDaoException;
-
+/**
+ * 
+ * @author Eduardo B , Yesid M
+ * Interfaz que contiene la estructura de los metos a implementar en la clase ClienteDAOHibernate
+ */
 public interface ClienteDAO {
 	public Cliente insertar(Cliente residente) throws IWDaoException;
 	public Cliente modificar(Cliente residente) throws IWDaoException;
 	public void eliminar(Cliente residente) throws IWDaoException;
 	public List <Cliente> obtener() throws IWDaoException;
 	public List<Cliente> obtenerByTipo(String tipo) throws IWDaoException;
-	public Cliente obtener(Long ficho) throws IWDaoException;
+	public Cliente obtener(Integer ficho) throws IWDaoException;
 }
