+++ src/co/udea/iw/dao/hibernate/ClienteDAOHibernate.java
@@ -13,9 +13,16 @@ import org.springframework.orm.hibernate3.support.HibernateDaoSupport;
 import co.udea.iw.dao.ClienteDAO;
 import co.udea.iw.dto.Cliente;
 import co.udea.iw.exception.IWDaoException;
-
+/**
+ * 
+ * @author Eduardo B , Yesid M
+ * Clase que implementa los metodos de la logica del negocio para la tabla de clientes
+ */
 public class ClienteDAOHibernate extends HibernateDaoSupport implements ClienteDAO {
 
+	/**
+	 * Metodo que permite insertar un cliente nuevo a la base de datos
+	 */
 	@Override
 	public Cliente insertar(Cliente cliente) throws IWDaoException {
 		Transaction tx = null;
@@ -32,7 +39,11 @@ public class ClienteDAOHibernate extends HibernateDaoSupport implements ClienteD
 		}
 		return cliente;
 	}
-
+	
+	
+     /**
+      * Metodo que permite modificar un cliente ya registrado en la base de datos 
+      */
 	@Override
 	public Cliente modificar(Cliente cliente) throws IWDaoException {
 		Transaction tx = null;
@@ -50,6 +61,9 @@ public class ClienteDAOHibernate extends HibernateDaoSupport implements ClienteD
 		return cliente;
 	}
 
+	/**
+	 * Metodo que permite eliminar un Cliente de la base de datos 
+	 */
 	@Override
 	public void eliminar(Cliente cliente) throws IWDaoException {
 		Transaction tx = null;
@@ -65,7 +79,12 @@ public class ClienteDAOHibernate extends HibernateDaoSupport implements ClienteD
 			session.close();
 		}
 	}
+	
+	
 
+	/**
+	 * Metodo que retorna una lista de clientes con todos los clientes regstrados en la base de datos 
+	 */
 	@Override
 	public List<Cliente> obtener() throws IWDaoException {
 		List<Cliente> clientes = new ArrayList<Cliente>();
@@ -80,7 +99,9 @@ public class ClienteDAOHibernate extends HibernateDaoSupport implements ClienteD
 	}
 	
 	
-	
+	/**
+	 * Metodo que retorna una lista de clientes por tipo de clinete registrado en la base de datos, el tipo el el parametro de entrada del metodo 
+	 */
 	@Override
 	public List<Cliente> obtenerByTipo(String tipo) throws IWDaoException {
 		List<Cliente> clientes = new ArrayList<>();
@@ -104,8 +125,11 @@ public class ClienteDAOHibernate extends HibernateDaoSupport implements ClienteD
 	}
 	
 
+	/**
+	 * Metodo que permite obtener un cliente con el numero de ficho del cliente
+	 */
 	@Override
-	public Cliente obtener(Long ficho) throws IWDaoException {
+	public Cliente obtener(Integer ficho) throws IWDaoException {
 		Cliente cliente = null;
 		Session session = null;
 		try {
